+++ src/main/resources/db/changelog/v0.8/add-old-events-geometries.sql
@@ -1,0 +1,5 @@
+--liquibase formatted sql
+
+--changeset event-api-migrations:v0.8/add-old-events-geometries.sql runOnChange:false
+
+update feed_data set geometries = collectEventGeometries(episodes) where geometries is null;